storage/requirements.txt
attrs==21.2.0
cached-property==1.5.2
certifi==2021.10.8
charset-normalizer==2.0.7
click==7.1.2
clickclick==20.10.2
connexion==2.9.0
DateTime==4.3
Flask==1.1.4
greenlet==1.1.2
idna==3.3
inflection==0.5.1
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
kazoo==2.5.0
logger==1.4
logging-config==1.0.4
MarkupSafe==2.0.1
mypy-extensions==0.4.3
mysql-connector-python==8.0.23
openapi-schema-validator==0.1.5
openapi-spec-validator==0.3.1
pykafka==2.8.0
PyMySQL==1.0.2
pyrsistent==0.18.0
pytz==2021.3
PyYAML==5.4.1
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.26
sqlalchemy-orm==1.2.2
swagger-ui-bundle==0.0.8
tabulate==0.8.9
typing-extensions==3.10.0.2
typing-inspect==0.7.1
urllib3==1.26.7
Werkzeug==1.0.1
zope.interface==5.4.0
